+++ packages/app/src/stores/editor.tsx
@@ -1,8 +1,7 @@
-import useSWR, { SWRResponse } from 'swr';
-import { useStaticSWR } from './use-static-swr';
-import { usePageId, useTemplateTagData, useShareLinkId } from '~/stores/context';
+import { SWRResponse } from 'swr';
 import { GetPageTagResponse } from '~/interfaces/tag';
-import { apiGet } from '~/client/util/apiv1-client';
+import { usePageTags } from '~/stores/context';
+import { useStaticSWR } from './use-static-swr';
 
 
 export const useIsSlackEnabled = (isEnabled?: boolean): SWRResponse<boolean, Error> => {
@@ -12,32 +11,7 @@ export const useIsSlackEnabled = (isEnabled?: boolean): SWRResponse<boolean, Err
   );
 };
 
-export const usePageTags = (): SWRResponse<string[] | undefined, Error> => {
-  const { data: pageId } = usePageId();
-  const { data: templateTagData } = useTemplateTagData();
-  const { data: shareLinkId } = useShareLinkId();
-
-  const fetcher = async(endpoint: string) => {
-    if (shareLinkId != null) {
-      return;
-    }
-
-    let tags: string[] = [];
-    // when the page exists or is a shared page
-    if (pageId != null && shareLinkId == null) {
-      const res = await apiGet<GetPageTagResponse>(endpoint, { pageId });
-      tags = res?.tags;
-    }
-    // when the page does not exist
-    else if (templateTagData != null) {
-      tags = templateTagData.split(',').filter((str: string) => {
-        return str !== ''; // filter empty values
-      });
-    }
-
-    return tags;
-  };
-
-  return useSWR('/pages.getPageTag', fetcher);
-
+export const useStaticPageTags = (): SWRResponse<GetPageTagResponse['tags'], Error> => {
+  const { data: pageId } = usePageTags();
+  return useStaticSWR<GetPageTagResponse['tags'], Error>('pageTags', pageId || null);
 };
