+++ packages/app/src/stores/page.tsx
@@ -6,6 +6,10 @@ import { HasObjectId } from '~/interfaces/has-object-id';
 import { IPage } from '~/interfaces/page';
 import { IPagingResult } from '~/interfaces/paging-result';
 
+import { usePageId, useTemplateTagData, useShareLinkId } from '~/stores/context';
+import { GetPageTagResponse } from '~/interfaces/tag';
+import { apiGet } from '~/client/util/apiv1-client';
+
 
 export const useSWRxPageByPath = (path: string, initialData?: IPage): SWRResponse<IPage & HasObjectId, Error> => {
   return useSWR(
@@ -42,4 +46,33 @@ export const useSWRxPageList = (
       };
     }),
   );
+};
+
+export const useStaticPageTags = (): SWRResponse<string[] | undefined, Error> => {
+  const { data: pageId } = usePageId();
+  const { data: templateTagData } = useTemplateTagData();
+  const { data: shareLinkId } = useShareLinkId();
+
+  const fetcher = async(endpoint: string) => {
+    if (shareLinkId != null) {
+      return;
+    }
+
+    let tags: string[] = [];
+    // when the page exists or is a shared page
+    if (pageId != null && shareLinkId == null) {
+      const res = await apiGet<GetPageTagResponse>(endpoint, { pageId });
+      tags = res?.tags;
+    }
+    // when the page does not exist
+    else if (templateTagData != null) {
+      tags = templateTagData.split(',').filter((str: string) => {
+        return str !== ''; // filter empty values
+      });
+    }
+
+    return tags;
+  };
+
+  return useSWR('/pages.getPageTag', fetcher);
 };
